Apply static casting to return value of Staff::getNumberOfLines().

+++ src/core/score/Staff.cpp
@@ -30,6 +30,6 @@ namespace raindrops
 
     int Staff::getNumberOfLines() const
     {
-        return m_staffLines.size();
+        return static_cast<int>(m_staffLines.size());
     }
 }
